requirements.txt
pandas
openai
genanki
python-dotenv
requests
pyunsplash
pyarrow